Fix for JRUBY-3381: File.rename does not report correct error codes
Because the Java APIs for renaming a file only report success or failure, we always just raise EACCES on failure. This breaks FileUtils.mv because it wants to rescue EXDEV in some cases to try a manual copy of the file. By adding a rescue of EACCES, we will try the copy logic, and if it's still an access failure, it will simply fail again the same way...but if it's EXDEV, it will work.