Attempt to find non-world-writable tmpdir
In JRuby, Dir.mktmpdir {} raises Errno::EACCES 'Permission denied'
because it calls FileUtils.remove_entry_secure internally to delete the
created tmpdir and remove_entry_secure tries to open the tmpdir for
reading before removing contents in the tmpdir.  Opening a directory for
reading is not allowed in Java.

FileUtils.remove_entry_secure tries to open the directory to avoid
TOCTOU file removing vulnerability.  When the parent directory of a
tmpdir is world-writable, FileUtils.remove_entry_secure needs to open
the tmpdir.  When the parent directory is NOT world-writable, it simply
calls FileUtils.remove_entry which works well with JRuby.  So this
commit changed Dir.tmpdir and attempts to find a directory which is not
world-writable first.  There's '.' in candidates and we can generally
expect it non-world-writable so Dir.mktmpdir {} should work.

Bear in mind that this change affects to Tempfile library (implemented
in Java for JRuby.) Files created by Tempfile.new is in Dir.tmpdir which
is changed by this commit.