See jruby/jruby#3663 and jruby/jruby@d1a760e06.
Use ipaddr to compare the sender's address
When resolving names using compressed IPv6-only DNS servers, it could
happen that when processing the response of a name query, the IP of the
sender is not compressed so using a simple string comparison is not
safe.

This patch encapsulates the from address and the address of the sender in
IPAddr objects so the comparison is safe and just works in all cases.

This fixes jruby/jruby#3663.